refactor(integrations): extract truenas connection handling into a client
